lib: fix MSVC compiler warnings
Visual C++ complained:
warning C4267: '=': conversion from 'size_t' to 'long', possible loss of data
warning C4701: potentially uninitialized local variable 'path' used

+++ lib/mprintf.c
@@ -935,7 +935,7 @@ static int dprintf_formatf(
           }
 
           if(prec > (long)maxprec)
-            prec = maxprec-1;
+            prec = (long)maxprec-1;
           /* RECURSIVE USAGE */
           len = curl_msnprintf(fptr, left, ".%ld", prec);
           fptr += len;
